@@ -60,31 +57,31 @@
     "release": "sh scripts/release/release.sh",
     "affected": "./node_modules/.bin/nx affected",
     "nx": "nx",
-    "postinstall": "ngcc --properties es2015 browser module main --first-only --async false && node ./decorate-angular-cli.js",
+    "postinstall": "ngcc --properties es2020 browser module main --first-only --async false && node ./decorate-angular-cli.js",
     "workspace-generator": "nx workspace-generator"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
-    "@angular/cdk": "~13.2.0",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/elements": "~13.2.0",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.2.0",
-    "@angular/material": "~13.2.0",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/router-store": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@nrwl/angular": "13.8.5",
-    "@ungap/custom-elements": "^1.0.0",
-    "d3": "^7.3.0",
-    "openseadragon": "^3.0.0",
-    "rxjs": "~7.4.0",
+    "@angular/animations": "14.2.2",
+    "@angular/cdk": "14.2.2",
+    "@angular/common": "14.2.2",
+    "@angular/compiler": "14.2.2",
+    "@angular/core": "14.2.2",
+    "@angular/elements": "14.2.2",
+    "@angular/flex-layout": "^14.0.0-beta.40",
+    "@angular/forms": "14.2.2",
+    "@angular/material": "14.2.2",
+    "@angular/platform-browser": "14.2.2",
+    "@angular/platform-browser-dynamic": "14.2.2",
+    "@angular/router": "14.2.2",
+    "@ngrx/effects": "14.3.1",
+    "@ngrx/router-store": "14.3.1",
+    "@ngrx/store": "14.3.1",
+    "@nrwl/angular": "14.7.8",
+    "@ungap/custom-elements": "^1.1.0",
+    "d3": "^7.6.1",
+    "openseadragon": "^3.1.0",
+    "rxjs": "~7.5.0",
     "stream": "^0.0.2",
     "timers": "^0.1.1",
     "tslib": "^2.0.0",
@@ -92,77 +89,84 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.0",
-    "@angular-eslint/builder": "~13.0.1",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/schematics": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.2.0",
-    "@angular/compiler-cli": "~13.2.0",
-    "@angular/language-service": "~13.2.0",
-    "@axe-core/webdriverjs": "^4.1.0",
+    "@angular-devkit/build-angular": "14.2.3",
+    "@angular-eslint/builder": "~14.0.4",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/schematics": "~14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.2.2",
+    "@angular/compiler-cli": "14.2.2",
+    "@angular/language-service": "14.2.2",
+    "@axe-core/playwright": "^4.4.4",
     "@babel/core": "^7.16.0",
     "@babel/preset-env": "^7.16.4",
-    "@ngrx/schematics": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "@nrwl/cli": "13.8.5",
-    "@nrwl/cypress": "13.8.5",
-    "@nrwl/eslint-plugin-nx": "13.8.5",
-    "@nrwl/jest": "13.8.5",
-    "@nrwl/linter": "13.8.5",
-    "@nrwl/nx-cloud": "13.1.6",
-    "@nrwl/tao": "13.8.5",
-    "@nrwl/workspace": "13.8.5",
+    "@cucumber/cucumber": "^8.5.3",
+    "@cucumber/html-formatter": "^20.1.0",
+    "@cucumber/messages": "^19.1.2",
+    "@cucumber/pretty-formatter": "^1.0.0",
+    "@ngrx/schematics": "14.0.2",
+    "@ngrx/store-devtools": "14.3.1",
+    "@nrwl/cli": "14.7.8",
+    "@nrwl/cypress": "14.7.8",
+    "@nrwl/eslint-plugin-nx": "14.7.8",
+    "@nrwl/jest": "14.7.8",
+    "@nrwl/linter": "14.7.8",
+    "@nrwl/nx-cloud": "14.6.2",
+    "@nrwl/workspace": "14.7.8",
+    "@playwright/test": "1.24.2",
     "@types/chai": "^4.2.5",
     "@types/chai-as-promised": "^7.1.2",
-    "@types/cucumber": "^6.0.0",
+    "@types/cucumber": "^7.0.0",
     "@types/d3": "^6.3.0",
-    "@types/jasmine": "~3.10.3",
+    "@types/jasmine": "4.0.3",
     "@types/jasminewd2": "~2.0.10",
-    "@types/jest": "27.0.2",
+    "@types/jest": "28.1.8",
+    "@types/jest-expect-message": "^1.0.4",
     "@types/node": "16.11.7",
     "@types/openseadragon": "^2.4.5",
     "@types/xml2js": "^0.4.8",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~5.10.0",
+    "@typescript-eslint/eslint-plugin": "5.33.1",
+    "@typescript-eslint/parser": "5.33.1",
+    "angular-http-server": "^1.11.1",
     "babel-loader": "^8.2.3",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "concurrently": "^5.0.0",
-    "cucumber": "^6.0.5",
+    "concurrently": "^7.4.0",
+    "cucumber-html-reporter": "^5.5.0",
     "cypress": "^9.1.0",
     "dotenv": "8.2.0",
-    "eslint": "~8.7.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-protractor": "^2.1.1",
-    "jasmine-core": "^3.10.1",
+    "eslint-plugin-playwright": "^0.11.1",
+    "jasmine-core": "4.2.0",
     "jasmine-marbles": "^0.9.2",
-    "jasmine-spec-reporter": "~7.0.0",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.1.1",
-    "karma": "6.3.17",
+    "jasmine-spec-reporter": "7.0.0",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-expect-message": "^1.1.2",
+    "jest-preset-angular": "12.2.2",
+    "karma": "6.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "^2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "1.7.0",
-    "mountebank": "^2.6.0",
-    "ng-packagr": "13.2.1",
-    "ngx-build-plus": "^13.0.1",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.0.0",
+    "mountebank": "^2.7.0",
+    "ng-packagr": "14.2.1",
+    "ngx-build-plus": "^14.0.0",
+    "nx": "14.7.8",
+    "playwright": "1.24.2",
     "postcss": "^8.3.9",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "2.5.1",
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
+    "prettier": "2.7.1",
     "pretty-quick": "^3.1.3",
-    "protractor": "7.0.0",
-    "protractor-cucumber-framework": "^6.2.0",
-    "protractor-multiple-cucumber-html-reporter-plugin": "^1.8.1",
     "selenium-webdriver": "^4.1.0",
     "standard-version": "^9.3.2",
-    "ts-jest": "27.0.5",
-    "ts-node": "9.1.1",
-    "typescript": "~4.5.2"
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.7.4"
   }
 }
